refactoring metrics
