+++ config/train_config.yaml
@@ -16,6 +16,7 @@ training_settings:
   save_total_limit: 1
   class_weighting: False # if True a custom loss function will be used which aims to deal with label imbalance
   output_weighting: False # if True custom thresholds will be set for the logits if False threshold = 0.5 for all
+  output_thresholds: [0.2, 0.5, 0.8, 0.95] # custom thresholds for top 4 predictions when output_weighting is True
 
 wandb_settings:
   project_name: 'grant_hrcs_tagger'
