+++ src/train.py
@@ -1,12 +1,10 @@
 import argparse
-import json
 import os
 
 import numpy as np
 import pandas as pd
 import torch
 import torch.nn as nn
-from sklearn.metrics import f1_score, precision_score, recall_score
 from transformers import (
     AutoModelForSequenceClassification,
     AutoTokenizer,
@@ -19,6 +17,7 @@ from transformers import (
 
 import wandb
 from utils import load_yaml_config
+from metrics import prepare_compute_metrics, plot_metrics, load_label_names
 
 
 def init_wandb(project_name, config_settings, report_to):
@@ -57,22 +56,6 @@ def init_device():
         return "cuda"
     else:
         return "cpu"
-
-def load_label_names(label_names_path):
-    """Load label names from a JSONL file.
-
-    Args:
-        label_names_path (str): Path to the label names file.
-
-    Returns:
-        dict: Dictionary mapping label IDs to label names.
-    """
-
-    with open(label_names_path) as f:
-        label_names = {k: v for line in f for k, v in json.loads(line).items()}
-        label_names = {idx: name for idx, name in enumerate(label_names.values())}
-
-    return label_names
 
 class WeightedTrainer(Trainer):
     def __init__(self, *args, class_weights: torch.FloatTensor | None = None, **kwargs):
@@ -163,32 +146,13 @@ def train(train_data, test_data, model_path, config, class_counts, class_weighti
     label_names = load_label_names(label_names_path)
 
     # initialize model
+    model = get_model(
+        config["training_settings"]["model"],
+        num_labels,
+        label_names
+    )
     if "modernbert" in config["training_settings"]["model"].lower():
-        model = ModernBertForSequenceClassification.from_pretrained(
-            config["training_settings"]["model"],
-            num_labels=num_labels,
-            problem_type="multi_label_classification",
-            id2label=label_names,
-            reference_compile=False,
-        )
-
-        print("model initialized using ModernBertForSequenceClassification")
-    elif "distilbert" in config["training_settings"]["model"].lower():
-        model = DistilBertForSequenceClassification.from_pretrained(
-            config["training_settings"]["model"],
-            num_labels=num_labels,
-            id2label=label_names,
-            problem_type="multi_label_classification",
-        )
-        print("model initialized using DistilBertForSequenceClassification")
-    else:
-        model = AutoModelForSequenceClassification.from_pretrained(
-            config["training_settings"]["model"],
-            num_labels=num_labels,
-            id2label=label_names,
-            problem_type="multi_label_classification",
-        )
-        print("model initialized using AutoModelForSequenceClassification")
+        model.config.reference_compile=False
 
     # initialize training arguments
     report_to_value = config["training_settings"]["report_to"]
@@ -249,106 +213,32 @@ def train(train_data, test_data, model_path, config, class_counts, class_weighti
     return metrics
 
 
-def prepare_compute_metrics(config):
-    """Wrapper for compute_metrics so config can be accessed
-
-    Args: config (dict): Configuration dictionary from yaml file.
-
-    Returns:
-        function: compute_metrics function
+def get_model(model_name, num_labels, label_names):
     """
-
-    def compute_metrics(eval_pred):
-        """Compute evaluation metrics to be used in the Trainer.
-
-        Args: eval_pred (tuple): Tuple containing logits and labels.
-
-        Returns:
-            dict: Evaluation metrics (f1, f1_macro, f1_micro, precision,
-            recall)
-        """
-        logits, labels = eval_pred
-        # apply sigmoid to logits
-        logits = torch.sigmoid(torch.tensor(logits)).cpu().detach().numpy()
-
-        num_tags_predicted = []
-        if config["training_settings"]["output_weighting"]:
-            thresholds = [1] * labels.shape[1]
-            thresholds[0:4] = [0.2, 0.5, 0.8, 0.95]
-
-            # Prepare an array to hold your predictions
-            predictions = np.zeros_like(logits)
-
-            # Loop through each sample's logits
-            for i, logit in enumerate(logits):
-                # Get the indices of the logits sorted by value in descending order
-                sorted_indices = np.argsort(logit)[::-1]
-
-                # Assign 1 to the top logits that exceed their respective thresholds
-                for rank, idx in enumerate(sorted_indices):
-                    if logit[idx] > thresholds[rank]:
-                        predictions[i, idx] = 1
-                num_tags_predicted.append(np.sum(predictions[i]))
-        else:
-            predictions = np.where(logits > 0.5, 1, 0)
-            for prediction in predictions:
-                num_tags_predicted.append(np.sum(prediction))
-
-            print(
-                "num_tags_predicted: ",
-                pd.Series(num_tags_predicted).value_counts().sort_index(),
-            )
-            log_data = pd.Series(num_tags_predicted).value_counts()
-            log_to_wandb({"num_tags_predicted": log_data.sort_index().to_json()}, config["training_settings"]["report_to"])
-
-        # compute actual metrics
-        f1_macro = f1_score(labels, predictions, average="macro")
-        f1_micro = f1_score(labels, predictions, average="micro")
-        f1 = f1_score(labels, predictions, average=None)
-        precision = precision_score(labels, predictions, average=None)
-        recall = recall_score(labels, predictions, average=None)
-        metrics = {
-            "f1": f1,
-            "f1_macro": f1_macro,
-            "f1_micro": f1_micro,
-            "precision": precision,
-            "recall": recall,
-        }
-        return metrics
-
-    return compute_metrics
-
-
-def plot_metrics(metrics, class_labels, train_counts, test_counts, config):
-    """Plot evaluation metrics and value counts in wandb
+    Loads a pre-trained model for sequence classification.
 
     Args:
-        metrics (dict): Evaluation metrics.
-        class_labels (list): List of class labels.
-        train_counts (list): List of class counts for train dataset.
-        test_counts (list): List of class counts for test dataset.
-        config (dict): Configuration dictionary.
+        model_name (str): The name of the model to load.
+        num_labels (int): The number of labels for classification.
+        label_names (dict): A dictionary mapping label IDs to label names.
 
+    Returns:
+        transformers.PreTrainedModel: The loaded model.
     """
+    model_class = AutoModelForSequenceClassification
+    if "modernbert" in model_name.lower():
+        model_class = ModernBertForSequenceClassification
+    elif "distilbert" in model_name.lower():
+        model_class = DistilBertForSequenceClassification
 
-    f1 = metrics["eval_f1"]
-    precision = metrics["eval_precision"]
-    recall = metrics["eval_recall"]
-    data = zip(
-        class_labels, precision, recall, f1, train_counts, test_counts, strict=False
+    model = model_class.from_pretrained(
+        model_name,
+        num_labels=num_labels,
+        problem_type="multi_label_classification",
+        id2label=label_names,
     )
-
-    # print data as a dataframe to console
-    df = pd.DataFrame(
-        list(data),
-        columns=["label", "precision", "recall", "f1", "train_count", "test_count"],
-    )
-    print(df)
-
-    # Only proceed if wandb reporting is enabled
-    if config["training_settings"]["report_to"] == "wandb":
-        table = wandb.Table(dataframe=df)
-        wandb.log({"metrics and value_count table": table})
+    print(f"Model initialized using {model_class.__name__}")
+    return model
 
 
 def run_training(args):
